+++ dash/dependencies.py
@@ -36,6 +36,9 @@ class ClientsideFunction:
     # pylint: disable=too-few-public-methods
     def __init__(self, namespace=None, function_name=None):
 
+        if namespace.startswith('_dashprivate_'):
+            raise ValueError("Namespaces cannot start with '_dashprivate_'.")
+
         if namespace in ['PreventUpdate', 'no_update']:
             raise ValueError('"{}" is a forbidden namespace in'
                              ' dash_clientside.'.format(namespace))